chore(react-tree): removes BaseTreeItem premature generalization
