+++ packages/react-components/react-tree/etc/react-tree.api.md
@@ -17,29 +17,6 @@ import * as React_2 from 'react';
 import type { Slot } from '@fluentui/react-utilities';
 import type { SlotClassNames } from '@fluentui/react-utilities';
 
-// @public
-export const BaseTreeItem: ForwardRefComponent<BaseTreeItemProps>;
-
-// @public (undocumented)
-export const baseTreeItemClassNames: SlotClassNames<BaseTreeItemSlots>;
-
-// @public
-export type BaseTreeItemProps = ComponentProps<BaseTreeItemSlots>;
-
-// @public (undocumented)
-export type BaseTreeItemSlots = {
-    root: Slot<'div'>;
-};
-
-// @public
-export type BaseTreeItemState = ComponentState<BaseTreeItemSlots> & {
-    open: boolean;
-    isLeaf: boolean;
-};
-
-// @public
-export const renderBaseTreeItem_unstable: (state: BaseTreeItemState) => JSX.Element;
-
 // @public (undocumented)
 export const renderTree_unstable: (state: TreeState, contextValues: TreeContextValues) => JSX.Element;
 
@@ -70,10 +47,11 @@ export const TreeItem: ForwardRefComponent<TreeItemProps>;
 export const treeItemClassNames: SlotClassNames<TreeItemSlots>;
 
 // @public
-export type TreeItemProps = ComponentProps<Partial<TreeItemSlots>> & BaseTreeItemProps;
+export type TreeItemProps = ComponentProps<Partial<TreeItemSlots>>;
 
 // @public (undocumented)
-export type TreeItemSlots = BaseTreeItemSlots & {
+export type TreeItemSlots = {
+    root: Slot<'div'>;
     expandIcon?: Slot<'span'>;
     iconBefore?: Slot<'span'>;
     iconAfter?: Slot<'span'>;
@@ -83,7 +61,9 @@ export type TreeItemSlots = BaseTreeItemSlots & {
 };
 
 // @public
-export type TreeItemState = ComponentState<TreeItemSlots> & BaseTreeItemState & {
+export type TreeItemState = ComponentState<TreeItemSlots> & {
+    open: boolean;
+    isLeaf: boolean;
     keepActionsOpen: boolean;
 };
 
@@ -108,12 +88,6 @@ export type TreeSlots = {
 export type TreeState = ComponentState<TreeSlots> & TreeContextValue & {
     open: boolean;
 };
-
-// @public
-export const useBaseTreeItem_unstable: (props: BaseTreeItemProps, ref: React_2.Ref<HTMLDivElement>) => BaseTreeItemState;
-
-// @public
-export const useBaseTreeItemStyles_unstable: (state: BaseTreeItemState) => BaseTreeItemState;
 
 // @public
 export const useTree_unstable: (props: TreeProps, ref: React_2.Ref<HTMLElement>) => TreeState;
