--- packages/react-components/react-tree/src/BaseTreeItem.ts
@@ -1,1 +1,0 @@
-export * from './components/BaseTreeItem/index';
