--- packages/react-components/react-tree/src/components/BaseTreeItem/index.ts
@@ -1,5 +1,0 @@
-export * from './BaseTreeItem';
-export * from './BaseTreeItem.types';
-export * from './renderBaseTreeItem';
-export * from './useBaseTreeItem';
-export * from './useBaseTreeItemStyles';
