+++ packages/react-components/react-tree/src/components/TreeItem/useTreeItem.tsx
@@ -1,15 +1,15 @@
 import * as React from 'react';
-import { isResolvedShorthand, resolveShorthand } from '@fluentui/react-utilities';
+import { getNativeElementProps, isResolvedShorthand, resolveShorthand } from '@fluentui/react-utilities';
 import { ChevronRightRegular } from '@fluentui/react-icons';
 import { useFluent_unstable } from '@fluentui/react-shared-contexts';
 import { useEventCallback } from '@fluentui/react-utilities';
 import { useFocusableGroup } from '@fluentui/react-tabster';
 import { expandIconInlineStyles } from './useTreeItemStyles';
-import { useBaseTreeItem_unstable } from '../BaseTreeItem/index';
-import { Enter } from '@fluentui/keyboard-keys';
+import { ArrowLeft, ArrowRight, Enter } from '@fluentui/keyboard-keys';
 import { useMergedRefs } from '@fluentui/react-utilities';
 import { elementContains } from '@fluentui/react-portal';
 import type { TreeItemProps, TreeItemState } from './TreeItem.types';
+import { useTreeContext_unstable } from '../../contexts/index';
 
 /**
  * Create the state required to render TreeItem.
@@ -21,30 +21,74 @@ import type { TreeItemProps, TreeItemState } from './TreeItem.types';
  * @param ref - reference to root HTMLElement of TreeItem
  */
 export const useTreeItem_unstable = (props: TreeItemProps, ref: React.Ref<HTMLDivElement>): TreeItemState => {
-  const treeItemState = useBaseTreeItem_unstable(props, ref);
+  const { 'aria-owns': ariaOwns, as = 'div', onClick, onKeyDown, ...rest } = props;
+  const level = useTreeContext_unstable(ctx => ctx.level);
+  const requestOpenChange = useTreeContext_unstable(ctx => ctx.requestOpenChange);
+  const focusFirstSubtreeItem = useTreeContext_unstable(ctx => ctx.focusFirstSubtreeItem);
+  const focusSubtreeOwnerItem = useTreeContext_unstable(ctx => ctx.focusSubtreeOwnerItem);
+
+  const isBranch = typeof ariaOwns === 'string';
+  const open = useTreeContext_unstable(ctx => isBranch && ctx.openSubtrees.includes(ariaOwns!));
   const { expandIcon, iconBefore, iconAfter, actions, badges, groupper } = props;
   const { dir, targetDocument } = useFluent_unstable();
-  const expandIconRotation = treeItemState.open ? 90 : dir !== 'rtl' ? 0 : 180;
+  const expandIconRotation = open ? 90 : dir !== 'rtl' ? 0 : 180;
   const groupperProps = useFocusableGroup();
 
   const actionsRef = React.useRef<HTMLElement>(null);
 
+  const handleArrowRight = (event: React.KeyboardEvent<HTMLDivElement>) => {
+    if (open && isBranch) {
+      focusFirstSubtreeItem(event.currentTarget);
+    }
+    if (isBranch && !open) {
+      requestOpenChange({ event, open: true, type: 'arrowRight', id: ariaOwns! });
+    }
+  };
+  const handleArrowLeft = (event: React.KeyboardEvent<HTMLDivElement>) => {
+    if (!isBranch || !open) {
+      focusSubtreeOwnerItem(event.currentTarget);
+    }
+    if (isBranch && open) {
+      requestOpenChange({ event, open: false, type: 'arrowLeft', id: ariaOwns! });
+    }
+  };
+  const handleEnter = (event: React.KeyboardEvent<HTMLDivElement>) => {
+    // if Enter keydown event comes from actions, ignore it
+    if (actionsRef.current && elementContains(actionsRef.current, event.target as Node)) {
+      return;
+    }
+    if (isBranch) {
+      requestOpenChange({ event, open: !open, type: 'enter', id: ariaOwns! });
+    }
+  };
+
   const handleClick = useEventCallback((event: React.MouseEvent<HTMLDivElement>) => {
+    onClick?.(event);
     //  if click event originates from actions, ignore it
     if (actionsRef.current && elementContains(actionsRef.current, event.target as Node)) {
       return;
     }
-    treeItemState.root.onClick?.(event);
+    if (isBranch) {
+      requestOpenChange({ event, open: !open, type: 'click', id: ariaOwns! });
+    }
   });
 
   const handleKeyDown = useEventCallback((event: React.KeyboardEvent<HTMLDivElement>) => {
-    if (event.key === Enter) {
-      // if Enter keydown event comes from actions, ignore it
-      if (actionsRef.current && elementContains(actionsRef.current, event.target as Node)) {
-        return;
+    onKeyDown?.(event);
+    if (event.isDefaultPrevented()) {
+      return;
+    }
+    switch (event.key) {
+      case Enter: {
+        return handleEnter(event);
+      }
+      case ArrowRight: {
+        return handleArrowRight(event);
+      }
+      case ArrowLeft: {
+        return handleArrowLeft(event);
       }
     }
-    treeItemState.root.onKeyDown?.(event);
   });
 
   const [keepActionsOpen, setKeepActionsOpen] = React.useState(false);
@@ -64,10 +108,11 @@ export const useTreeItem_unstable = (props: TreeItemProps, ref: React.Ref<HTMLDi
   }, [targetDocument]);
 
   return {
-    ...treeItemState,
+    isLeaf: !isBranch,
+    open,
     keepActionsOpen,
     components: {
-      ...treeItemState.components,
+      root: 'div',
       expandIcon: 'span',
       iconBefore: 'span',
       iconAfter: 'span',
@@ -75,11 +120,18 @@ export const useTreeItem_unstable = (props: TreeItemProps, ref: React.Ref<HTMLDi
       badges: 'span',
       groupper: 'span',
     },
-    root: {
-      ...treeItemState.root,
+    root: getNativeElementProps(as, {
+      ...rest,
+      ref,
+      tabIndex: 0,
+      'aria-owns': ariaOwns,
+      'aria-level': level,
+      // FIXME: tabster fails to navigate when aria-expanded is true
+      // 'aria-expanded': isBranch ? isOpen : undefined,
+      role: 'treeitem',
       onClick: handleClick,
       onKeyDown: handleKeyDown,
-    },
+    }),
     groupper: resolveShorthand(groupper, {
       required: true,
       defaultProps: {
@@ -94,7 +146,7 @@ export const useTreeItem_unstable = (props: TreeItemProps, ref: React.Ref<HTMLDi
       defaultProps: { 'aria-hidden': true },
     }),
     expandIcon: resolveShorthand(expandIcon, {
-      required: !treeItemState.isLeaf,
+      required: isBranch,
       defaultProps: {
         children: <ChevronRightRegular style={expandIconInlineStyles[expandIconRotation]} />,
         'aria-hidden': true,
