+++ packages/react-components/react-tree/src/index.ts
@@ -1,15 +1,6 @@
 export { Tree, treeClassNames, renderTree_unstable, useTreeStyles_unstable, useTree_unstable } from './Tree';
 export type { TreeProps, TreeState, TreeSlots } from './Tree';
 
-export {
-  BaseTreeItem,
-  baseTreeItemClassNames,
-  renderBaseTreeItem_unstable,
-  useBaseTreeItemStyles_unstable,
-  useBaseTreeItem_unstable,
-} from './BaseTreeItem';
-export type { BaseTreeItemProps, BaseTreeItemState, BaseTreeItemSlots } from './BaseTreeItem';
-
 export type { TreeContextValue, useTreeContext_unstable, TreeProvider } from './contexts';
 
 export {
